+++ src/main/kotlin/me/fzzyhmstrs/amethyst_imbuement/loot/MoStructuresLoot.kt
@@ -3,9 +3,9 @@ package me.fzzyhmstrs.amethyst_imbuement.loot
 import me.fzzyhmstrs.amethyst_core.item_util.AbstractModLoot
 import me.fzzyhmstrs.amethyst_imbuement.registry.RegisterItem
 import me.fzzyhmstrs.amethyst_imbuement.registry.RegisterLoot
-import net.fabricmc.fabric.api.loot.v1.FabricLootPoolBuilder
-import net.fabricmc.fabric.api.loot.v1.FabricLootSupplierBuilder
+import net.fabricmc.fabric.api.loot.v2.FabricLootTableBuilder
 import net.minecraft.item.Items
+import net.minecraft.loot.LootPool
 import net.minecraft.loot.entry.ItemEntry
 import net.minecraft.loot.provider.number.UniformLootNumberProvider
 import net.minecraft.util.Identifier
@@ -13,7 +13,7 @@ import net.minecraft.util.Identifier
 object MoStructuresLoot: AbstractModLoot() {
     override val targetNameSpace: String = "mostructures"
 
-    override fun lootBuilder(id: Identifier, table: FabricLootSupplierBuilder): Boolean{
+    override fun lootBuilder(id: Identifier, table: FabricLootTableBuilder): Boolean{
         when (id) {
             Identifier("mostructures","bunny_skeleton") -> {
                 VanillaLoot.shipwreckTreasureLoot(table)
@@ -25,15 +25,15 @@ object MoStructuresLoot: AbstractModLoot() {
             }
             Identifier("mostructures","ice_tower") -> {
                 val poolBuilder = RegisterLoot.tierOneGemPool(3.0F, 0.20F)
-                table.pool(poolBuilder)
+                table.pool(poolBuilder.build())
                 return true
             }
             Identifier("mostructures","barnhouse_3") -> {
-                val poolBuilder = FabricLootPoolBuilder.builder()
+                val poolBuilder = LootPool.builder()
                     .rolls(UniformLootNumberProvider.create(1.0F,4.0F))
                     .with(ItemEntry.builder(RegisterItem.XP_BUSH_SEED).weight(1))
                     .with(ItemEntry.builder(Items.AIR).weight(19))
-                table.pool(poolBuilder)
+                table.pool(poolBuilder.build())
                 return true
             }
             Identifier("mostructures","jungle_temple_treasure") -> {
@@ -46,9 +46,9 @@ object MoStructuresLoot: AbstractModLoot() {
             }
             Identifier("mostructures","pirate_ship_cargo_misc") -> {
                 val poolBuilder = RegisterLoot.tierOneGemPool(3.0F, 0.333F)
-                table.pool(poolBuilder)
+                table.pool(poolBuilder.build())
                 val poolBuilder2 = RegisterLoot.tierTwoGemPool(1.0F, 0.1F)
-                table.pool(poolBuilder2)
+                table.pool(poolBuilder2.build())
                 return true
             }
             Identifier("mostructures","pyramid_custom") -> {