+++ src/main/kotlin/me/fzzyhmstrs/amethyst_imbuement/registry/RegisterLoot.kt
@@ -2,8 +2,9 @@ package me.fzzyhmstrs.amethyst_imbuement.registry
 
 import me.fzzyhmstrs.amethyst_core.registry.LootRegistry.registerModLoot
 import me.fzzyhmstrs.amethyst_imbuement.loot.*
-import net.fabricmc.fabric.api.loot.v1.FabricLootPoolBuilder
+import net.fabricmc.fabric.api.loot.v2.FabricLootPoolBuilder
 import net.minecraft.item.Items
+import net.minecraft.loot.LootPool
 import net.minecraft.loot.entry.ItemEntry
 import net.minecraft.loot.provider.number.UniformLootNumberProvider
 import net.minecraft.util.math.MathHelper
@@ -23,10 +24,10 @@ object RegisterLoot {
         registerModLoot(BetterDungeonsLoot)
     }
 
-    fun tierOneGemPool(maxCount: Float, overallChance: Float): FabricLootPoolBuilder {
+    fun tierOneGemPool(maxCount: Float, overallChance: Float): LootPool.Builder {
         val actualChance = MathHelper.clamp(overallChance,0.01F,1.0F*maxCount)
         val airWeight = (50 * ((1.0F + maxCount) / 2) / actualChance) - 50
-        return FabricLootPoolBuilder.builder()
+        return LootPool.builder()
             .rolls(UniformLootNumberProvider.create(1.0F,maxCount))
             .with(ItemEntry.builder(RegisterItem.CITRINE).weight(10))
             .with(ItemEntry.builder(RegisterItem.DANBURITE).weight(10))
@@ -36,10 +37,10 @@ object RegisterLoot {
             .with(ItemEntry.builder(RegisterItem.GEM_DUST).weight(1))
             .with(ItemEntry.builder(Items.AIR).weight(max(airWeight.toInt(),1)))
     }
-    fun tierTwoGemPool(maxCount: Float, overallChance: Float): FabricLootPoolBuilder {
+    fun tierTwoGemPool(maxCount: Float, overallChance: Float): LootPool.Builder {
         val actualChance = MathHelper.clamp(overallChance,0.01F,1.0F*maxCount)
         val airWeight = (60 * ((1.0F + maxCount) / 2) / actualChance) - 60
-        return FabricLootPoolBuilder.builder()
+        return LootPool.builder()
             .rolls(UniformLootNumberProvider.create(1.0F,maxCount))
             .with(ItemEntry.builder(RegisterItem.MOONSTONE).weight(10))
             .with(ItemEntry.builder(RegisterItem.OPAL).weight(10))