+++ changelog.md
@@ -2,7 +2,7 @@
 * Added config option to not give the glistering tome on startup. The "scepters" config is now names "items"
 
 ### Changes
-* None
+* Update AC version and update loot table builders to new API version
 
 ### Fixes
 * Fixed summon totem of wit to be a wit spell